Fixing EggHatcher Errors

+++ NativePrograms/PokemonSV/EggHatcher.c
@@ -4,4 +4,4 @@
 const uint8_t BOXES_TO_HATCH    =   2;
 const uint16_t STEPS_TO_HATCH   =   5120;
 const bool HAS_CLONED_RIDER     =   false;
-const uint16_t SAFETY_COEFF     =   1.1;
+const float SAFETY_COEFF        =   1.1;
