+++ NativePrograms/PokemonSV/Programs/EggHatcher.h
@@ -3,4 +3,4 @@
 extern const uint8_t BOXES_TO_HATCH;
 extern const uint16_t STEPS_TO_HATCH;
 extern const bool HAS_CLONED_RIDER;
-extern const uint16_t SAFETY_COEFF;
+extern const float SAFETY_COEFF;
